+++ src/command.rs
@@ -80,7 +80,7 @@ pub enum Command {
 
 pub fn run(cmd: Option<Command>, game: &mut Game) -> Result<()> {
     match cmd.unwrap_or(Command::Stat) {
-        Command::Stat => log::status(&game),
+        Command::Stat => log::status(game),
         Command::ChangeDir {
             destination,
             run,
@@ -105,7 +105,7 @@ pub fn run(cmd: Option<Command>, game: &mut Game) -> Result<()> {
 /// Attempt to move the hero to the supplied location, possibly engaging
 /// in combat along the way.
 fn change_dir(game: &mut Game, dest: &str, run: bool, bribe: bool, force: bool) -> Result<()> {
-    let dest = Location::from(&dest)?;
+    let dest = Location::from(dest)?;
     if force {
         game.location = dest;
     } else if let Err(character::Dead) = game.go_to(&dest, run, bribe) {
@@ -160,7 +160,7 @@ fn shop(game: &mut Game, items: &[String]) -> Result<()> {
 /// Use an item from the inventory or list the inventory contents if no item name is provided.
 fn use_item(game: &mut Game, items: &[String]) -> Result<()> {
     if items.is_empty() {
-        println!("{}", log::format_inventory(&game));
+        println!("{}", log::format_inventory(game));
     } else {
         for item_name in items {
             let item_name = sanitize(item_name);
@@ -179,11 +179,19 @@ fn sanitize(name: &str) -> String {
         "es" => "escape",
         "sw" => "sword",
         "sh" => "shield",
-        "hp" | "health" => "hp-stone",
-        "mp" | "magic" => "mp-stone",
+        "hp" => "hp-stone",
+        "mp" => "mp-stone",
         "str" | "strength" => "str-stone",
         "spd" | "speed" => "spd-stone",
         "level" | "lv" | "lvl" => "lvl-stone",
+        "void" => "void-rng",
+        "att-ring" | "att" | "attack" | "attack-ring" | "attack-rng" => "att-rng",
+        "def-ring" | "def" | "deffense" | "deffense-ring" | "deffense-rng" => "def-rng",
+        "spd-ring" | "speed-ring" | "speed-rng" => "spd-rng",
+        "mag-ring" | "mag" | "magic-ring" | "magic-rng" => "mag-rng",
+        "mp-ring" => "mp-rng",
+        "hp-ring" => "hp-rng",
+        "evade" | "evade-ring" => "evade-rng",
         n => n,
     };
     name.to_string()
@@ -226,8 +234,7 @@ mod tests {
         };
 
         // reduce stats to ensure loss
-        game.player.speed = 1;
-        game.player.strength = 1;
+        // FIXME this is flaky now
         game.player.current_hp = 1;
 
         game.gold = 100;
@@ -237,7 +244,7 @@ mod tests {
 
         assert!(result.is_err());
         // game reset
-        assert_eq!(game.player.max_hp, game.player.current_hp);
+        assert_eq!(game.player.max_hp(), game.player.current_hp);
         assert_eq!(0, game.gold);
         assert_eq!(0, game.player.xp);
     }
@@ -273,7 +280,7 @@ mod tests {
         let result = run(Some(cmd), &mut game);
         assert!(result.is_ok());
         assert!(game.location.is_home());
-        assert_eq!(game.player.max_hp, game.player.current_hp);
+        assert_eq!(game.player.max_hp(), game.player.current_hp);
     }
 
     #[test]
@@ -290,8 +297,6 @@ mod tests {
         };
 
         // reduce stats to ensure loss
-        game.player.speed = 1;
-        game.player.strength = 1;
         game.player.current_hp = 1;
 
         game.gold = 100;
@@ -350,7 +355,7 @@ mod tests {
         let result = run(Some(cmd), &mut game);
         assert!(result.is_ok());
         assert!(game.inventory().is_empty());
-        assert_eq!(game.player.max_hp, game.player.current_hp);
+        assert_eq!(game.player.max_hp(), game.player.current_hp);
 
         // not buy if not home
         let cmd = Command::ChangeDir {