+++ src/item/equipment.rs
@@ -1,6 +1,7 @@
 use core::fmt;
 
-use crate::character::class as character;
+use super::ring::Ring;
+use crate::character::class::Class;
 use serde::{Deserialize, Serialize};
 
 /// Packages together different equipment pieces that can be worn by characters
@@ -8,10 +9,12 @@ use serde::{Deserialize, Serialize};
 /// Provides a unified interface for stat contributions to the base stats of a
 /// characters, e.g. the increased attack contributed by a sword and the
 /// deffense contributed by a shield.
-#[derive(Serialize, Deserialize, Debug, Default)]
+#[derive(Serialize, Deserialize, Default)]
 pub struct Equipment {
     pub sword: Option<Weapon>,
     pub shield: Option<Weapon>,
+    pub left_ring: Option<Ring>,
+    pub right_ring: Option<Ring>,
 }
 
 impl Equipment {
@@ -19,6 +22,8 @@ impl Equipment {
         Self {
             sword: None,
             shield: None,
+            left_ring: None,
+            right_ring: None,
         }
     }
 
@@ -32,12 +37,63 @@ impl Equipment {
         )
     }
 
-    pub fn attack(&self) -> i32 {
-        self.sword.as_ref().map_or(0, |s| s.strength())
+    /// Set the given ring as the left ring.
+    /// If there was a left ring already, put it in the right.
+    /// If there was a right ring already, return it.
+    pub fn put_ring(&mut self, ring: Ring) -> Option<Ring> {
+        let removed = self.right_ring.take();
+        self.right_ring = self.left_ring.take();
+        self.left_ring = Some(ring);
+        removed
     }
 
-    pub fn deffense(&self) -> i32 {
+    pub fn remove_ring(&mut self, name: &str) -> Option<Ring> {
+        match (&self.left_ring, &self.right_ring) {
+            (Some(ring), _) if ring.key() == name => {
+                let removed = self.left_ring.take();
+                self.left_ring = self.right_ring.take();
+                removed
+            }
+            (_, Some(ring)) if ring.key() == name => self.right_ring.take(),
+            _ => None,
+        }
+    }
+
+    pub fn attack(&self, strength: i32) -> i32 {
+        self.sword.as_ref().map_or(0, |s| s.strength())
+            + self.ring_contribution(strength, Ring::Attack)
+    }
+
+    pub fn deffense(&self, strength: i32) -> i32 {
         self.shield.as_ref().map_or(0, |s| s.strength())
+            + self.ring_contribution(strength, Ring::Deffense)
+    }
+
+    pub fn speed(&self, base: i32) -> i32 {
+        self.ring_contribution(base, Ring::Speed)
+    }
+
+    pub fn magic(&self, base: i32) -> i32 {
+        self.ring_contribution(base, Ring::Magic)
+    }
+
+    pub fn mp(&self, base: i32) -> i32 {
+        self.ring_contribution(base, Ring::MP)
+    }
+
+    pub fn hp(&self, base: i32) -> i32 {
+        self.ring_contribution(base, Ring::HP)
+    }
+
+    pub fn enemies_evaded(&self) -> bool {
+        self.left_ring == Some(Ring::Evade) || self.right_ring == Some(Ring::Evade)
+    }
+
+    fn ring_contribution(&self, base: i32, ring: Ring) -> i32 {
+        let factor =
+            |r: &Option<Ring>| r.as_ref().filter(|&l| *l == ring).map_or(0.0, Ring::factor);
+        let factor = factor(&self.left_ring) + factor(&self.right_ring);
+        (base as f64 * factor).round() as i32
     }
 }
 
@@ -71,7 +127,7 @@ impl Weapon {
     /// the item is equipped.
     pub fn strength(&self) -> i32 {
         // get the base strength of the hero at this level
-        let player_strength = character::Class::player_first().strength.at(self.level());
+        let player_strength = Class::player_first().strength.at(self.level());
 
         // calculate the added strength as a function of the player strength
         (player_strength as f64 * 0.5).round() as i32
@@ -94,5 +150,25 @@ impl fmt::Display for Weapon {
             Weapon::Shield(_) => "shield",
         };
         write!(f, "{}[{}]", name, self.level())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_apply_factor() {
+        let mut equip = Equipment::new();
+        assert_eq!(0, equip.ring_contribution(10, Ring::HP));
+        equip.left_ring = Some(Ring::Void);
+        equip.right_ring = Some(Ring::Void);
+        assert_eq!(0, equip.ring_contribution(10, Ring::HP));
+
+        equip.left_ring = Some(Ring::HP);
+        assert_eq!(5, equip.ring_contribution(10, Ring::HP));
+
+        equip.right_ring = Some(Ring::HP);
+        assert_eq!(10, equip.ring_contribution(10, Ring::HP));
     }
 }