+++ src/log.rs
@@ -56,7 +56,7 @@ pub fn handle(game: &Game, event: &Event) {
             items,
             ..
         } => {
-            battle_won(&game, *xp, *levels_up, *gold, items);
+            battle_won(game, *xp, *levels_up, *gold, items);
         }
         Event::BattleLost => {
             battle_lost(&game.player);
@@ -146,7 +146,7 @@ pub fn quest_list(quests: Vec<(bool, String)>) {
 
 pub fn quest_done(reward: i32) {
     if !quiet() {
-        println!("    {} quest completed!", format_gold_plus(reward));
+        println!("   {} quest completed!", format_gold_plus(reward));
     }
 }
 
@@ -236,14 +236,14 @@ fn attack(character: &Character, attack: &AttackType, damage: i32, mp_cost: i32)
     if !quiet() {
         battle_log(
             character,
-            &format_attack(character, &attack, damage, mp_cost),
+            &format_attack(character, attack, damage, mp_cost),
         );
     }
 }
 
 fn status_effect_damage(character: &Character, damage: i32) {
     let (_, emoji) = status_effect_params(character.status_effect.unwrap());
-    battle_log(character, &format_damage(character, damage, &emoji));
+    battle_log(character, &format_damage(character, damage, emoji));
 }
 
 fn battle_lost(player: &Character) {
@@ -278,7 +278,7 @@ fn stat_increase(player: &Character, stat: &str, increase: i32) {
     } else {
         format!("+{}{}", increase, stat).cyan().to_string()
     };
-    battle_log(&player, &suffix);
+    battle_log(player, &suffix);
 }
 
 fn long_status(game: &Game) {
@@ -290,11 +290,11 @@ fn long_status(game: &Game) {
         "    hp:{} {}/{}",
         hp_display(player, 10),
         player.current_hp,
-        player.max_hp
+        player.max_hp()
     );
 
     let (current_mp, max_mp) = if player.class.is_magic() {
-        (player.current_mp, player.max_mp)
+        (player.current_mp, player.max_mp())
     } else {
         (0, 0)
     };
@@ -319,7 +319,7 @@ fn long_status(game: &Game) {
         player.physical_attack(),
         player.magic_attack(),
         player.deffense(),
-        player.speed
+        player.speed()
     );
     println!("    {}", format_equipment(player));
     println!("    {}", format_inventory(game));
@@ -335,7 +335,7 @@ fn short_status(game: &Game) {
     } else {
         ""
     };
-    log(player, &game.location, &suffix);
+    log(player, &game.location, suffix);
 }
 
 fn plain_status(game: &Game) {
@@ -354,15 +354,15 @@ fn plain_status(game: &Game) {
         player.level,
         game.location,
         player.current_hp,
-        player.max_hp,
+        player.max_hp(),
         player.current_mp,
-        player.max_mp,
+        player.max_mp(),
         player.xp,
         player.xp_for_next(),
         player.magic_attack(),
         player.physical_attack(),
         player.deffense(),
-        player.speed,
+        player.speed(),
         status_effect,
         format_equipment(player),
         format_inventory(game),
@@ -437,6 +437,15 @@ fn format_equipment(character: &Character) -> String {
     if let Some(shield) = &character.equip.shield {
         fragments.push(shield.to_string());
     }
+
+    if let Some(ring) = &character.equip.left_ring {
+        fragments.push(ring.to_string());
+    }
+
+    if let Some(ring) = &character.equip.right_ring {
+        fragments.push(ring.to_string());
+    }
+
     format!("equip:{{{}}}", fragments.join(","))
 }
 
@@ -455,7 +464,7 @@ fn format_attack(receiver: &Character, attack: &AttackType, damage: i32, mp_cost
     let magic_effect = if mp_cost > 0 { "\u{2728}" } else { "" };
 
     match attack {
-        AttackType::Regular => format_damage(receiver, damage, &magic_effect),
+        AttackType::Regular => format_damage(receiver, damage, magic_effect),
         AttackType::Critical => {
             format_damage(receiver, damage, &format!("{} critical!", magic_effect))
         }
@@ -491,7 +500,7 @@ fn hp_display(character: &Character, slots: i32) -> String {
     bar_display(
         slots,
         character.current_hp,
-        character.max_hp,
+        character.max_hp(),
         "green",
         "red",
     )
@@ -507,7 +516,7 @@ fn mp_display(character: &Character, slots: i32) -> String {
     bar_display(
         slots,
         current_mp,
-        character.max_mp,
+        character.max_mp(),
         "purple",
         "bright black",
     )